boot: change bootloader copy from byte to word transfers

+++ abstract-machine/am/src/riscv/ysyxsoc/bootloader.c
@@ -1,28 +1,28 @@
 
 #include <klib.h>
 
-extern uint8_t _start;
+extern uint32_t _start;
 
-extern uint8_t _text_start;
-extern uint8_t _text_end;
-extern uint8_t _text_lma;
-extern uint8_t _rodata_start;
+extern uint32_t _text_start;
+extern uint32_t _text_end;
+extern uint32_t _text_lma;
+extern uint32_t _rodata_start;
 
-extern uint8_t _data_end;
-extern uint8_t _data_lma;
-extern uint8_t _data_start;
+extern uint32_t _data_end;
+extern uint32_t _data_lma;
+extern uint32_t _data_start;
 
-extern uint8_t _rodata_end;
-extern uint8_t _rodata_lma;
+extern uint32_t _rodata_end;
+extern uint32_t _rodata_lma;
 
-extern uint8_t _bss_start;
-extern uint8_t _bss_end;
-extern uint8_t _bss_lma;
+extern uint32_t _bss_start;
+extern uint32_t _bss_end;
+extern uint32_t _bss_lma;
 
 
 
 __attribute__((section(".boot.text"), noinline))
-static void boot_copy(uint8_t *dst, const uint8_t *src, size_t size)
+static void boot_copy(uint32_t *dst, const uint32_t *src, size_t size)
 {
     while (size != 0) {
         *dst++ = *src++;
@@ -31,7 +31,7 @@ static void boot_copy(uint8_t *dst, const uint8_t *src, size_t size)
 }
 
 __attribute__((section(".boot.text"), noinline))
-static void boot_zero(uint8_t *dst, size_t size)
+static void boot_zero(uint32_t *dst, size_t size)
 {
     while (size != 0) {
         *dst++ = 0;
